requirements.txt
args==0.1.0
certifi==2019.9.11
chardet==3.0.4
clint==0.5.1
FMPy==0.2.13
idna==2.8
joblib==0.13.2
lxml==4.4.1
numpy==1.17.2
pandas==0.25.1
pathlib==1.0.1
python-dateutil==2.8.0
pytz==2019.2
pywin32==225
requests==2.22.0
scikit-learn==0.21.3
scipy==1.3.1
six==1.12.0
SQLAlchemy==1.3.8
urllib3==1.25.6
wincertstore==0.2
matplotlib
pymysql
pyyaml
hbmqtt